+++ src/ast/passes/named_param.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "ast/pass_manager.h"
+#include "globalvars.h"
+#include "types.h"
+
+namespace bpftrace::ast {
+
+class NamedParamDefaults : public ast::State<"named_params_defaults"> {
+public:
+  std::unordered_map<std::string, globalvars::GlobalVarValue> defaults;
+};
+
+Pass CreateNamedParamsPass();
+
+} // namespace bpftrace::ast